Make line intersections exclude endpoints.
That is, if checking for an intersection between two line segments,
where there is an intersection, if the intersection is the endpoint of
one of the lines, we don't count it.

This fixes a minor bug where you could stand with your right side flush
against the left side of a bouncy platform (or vice versa) and jump
straight up, and when you came back down your BR corner would hit the TL
of the platform and so you would bounce, even though you haven't moved
any more over the platform than when you left the ground.