Add the original license from the sound code.
It comes from titanion, by Kenta Cho (ABA games). The source for that
game is under the BSD license, so I believe it is okay to use this code
(with or without modifications) as long as we keep the license text with
it. Ideally, we should rewrite the sound code so that it's actually our
own, but for now this will work.

Fix #40.